Main refactor completed.
I have refactored out the node stuff which defines a Ringo. This allows for a clear separation between the things which control a ringo, and the things which define a ringo. By doing this, we remove our reliance on using immutable String[] to represent a ringo, and can instead use dynamic POJO's. This provides both cleaner code and additional flexibility for future changes.

-Alex